pyproject.toml
[project]
name = "desktopmate-plus-backend"
version = "0.1.0"
description = "AI-powered desktop companion backend with vision, speech, and memory capabilities"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Core Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "python-multipart>=0.0.12",
    "websockets>=14.0",
    # LangGraph & LangChain
    "langgraph>=0.6.10",
    "langchain-openai>=0.3.33",
    "langchain-core>=0.3.0",
    "langchain-neo4j>=0.5.0",
    # LLM & AI Model Clients
    "openai>=1.0.0",
    # Memory Management
    "mem0ai>=1.0.0",
    "langmem>=0.0.29",
    # Database & Storage
    "psycopg[binary]>=3.2.1",
    "psycopg2-binary>=2.9.11",
    "qdrant-client>=1.15.1",
    "pymongo>=4.15.3",
    "rank-bm25>=0.2.2",
    # Screen Capture
    "mss>=9.0.0",
    "pillow>=11.0.0",
    # HTTP Client for external API calls
    "httpx>=0.27.0",
    "aiohttp>=3.11.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.6.0",
    "loguru>=0.7.0",
    "tenacity>=8.0.0",
    "types-requests>=2.32.4.20250913",
    "pyyaml>=6.0.0",
    "langchain-mcp-adapters>=0.1.11",
    "langchain>=0.3.27",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.27.0",
    "websockets>=14.0",

    # Code Quality
    "ruff>=0.8.0",
    "black>=24.10.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",

    # Development
    "jupyter>=1.1.1",
    "notebook>=7.4.7",
    "ipython>=8.30.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ["py313"]
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # function calls in argument defaults (FastAPI Depends pattern)
]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
